+++ tests/tests_unit/test_cdf_tk/test_validation/test_read_yaml.py
@@ -8,6 +8,7 @@ from cognite.client.data_classes.data_modeling import ViewApply
 
 from cognite_toolkit._cdf_tk._parameters import ParameterSpecSet
 from cognite_toolkit._cdf_tk.load import ContainerLoader, SpaceLoader
+from cognite_toolkit._cdf_tk.user_warnings import ToolkitWarning
 from cognite_toolkit._cdf_tk.validation import validate_case_raw, validate_data_set_is_set, validate_yaml_config
 from cognite_toolkit._cdf_tk.validation.warning import (
     CaseTypoWarning,
@@ -16,7 +17,6 @@ from cognite_toolkit._cdf_tk.validation.warning import (
     SnakeCaseWarning,
     UnusedParameterWarning,
 )
-from cognite_toolkit._cdf_tk.warnings import ToolkitWarning
 from tests.tests_unit.data import LOAD_DATA
 
 DUMMY_FILE = Path("dummy.yaml")
